refactor split AgdaToScala - extract Terms and Types
